backend/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
jinja2
pydantic[email]
# hash password
passlib[bcrypt]
argon2_cffi
# session
itsdangerous